Version1.1

+++ WalkingTours/SearchService.m
@@ -25,7 +25,7 @@
         if (searchResult.location == NSNotFound) {
             NSLog(@"Search term string was not found");
         } else {
-            NSLog(@"%@ starts at index %lu and is %lu characters long", lowerCaseSearchTerm, searchResult.location, searchResult.length);
+            NSLog(@"%@ starts at index %lu and is %lu characters long", lowerCaseSearchTerm, (unsigned long)searchResult.location, (unsigned long)searchResult.length);
             [matches addObject:[arrayToSearch objectAtIndex:i]];
         }
     }
